# RecruitmentAgency/RecruitmentAgency.WASM/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RecruitmentAgency.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Employer": {
      "get": {
        "tags": [
          "Employer"
        ],
        "summary": "Возвращает список всех работодателей.",
        "responses": {
          "200": {
            "description": "Список успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employer"
        ],
        "summary": "Добавляет нового работодателя.",
        "requestBody": {
          "description": "Информация о новом работодателе.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Работодатель успешно добавлен.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              }
            }
          },
          "400": {
            "description": "Работодатель не был добавлен."
          }
        }
      },
      "put": {
        "tags": [
          "Employer"
        ],
        "summary": "Обновляет информацию о существующем работодателе.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор работодателя",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновлённая информация о работодателе.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Работодатель успешно обновлён.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              }
            }
          },
          "404": {
            "description": "Работодатель с указанным идентификатором не найден."
          }
        }
      }
    },
    "/api/Employer/{id}": {
      "get": {
        "tags": [
          "Employer"
        ],
        "summary": "Получает информацию о работодателе по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор работодателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Работодатель найден и возвращён успешно.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerDTO"
                }
              }
            }
          },
          "404": {
            "description": "Работодатель с указанным идентификатором не найден."
          }
        }
      },
      "delete": {
        "tags": [
          "Employer"
        ],
        "summary": "Удаляет работодателя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор работодателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Работодатель успешно удалён."
          },
          "404": {
            "description": "Работодатель с указанным идентификатором не найден."
          }
        }
      }
    },
    "/api/EmployerApplication": {
      "get": {
        "tags": [
          "EmployerApplication"
        ],
        "summary": "Возвращает список всех заявок работодателей.",
        "responses": {
          "200": {
            "description": "Список успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerApplicationDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerApplicationDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployerApplicationDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EmployerApplication"
        ],
        "summary": "Добавляет новую заявку работодателя.",
        "requestBody": {
          "description": "Информация о новой заявке.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заявка успешно добавлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              }
            }
          },
          "400": {
            "description": "Заявка не была добавлена."
          }
        }
      },
      "put": {
        "tags": [
          "EmployerApplication"
        ],
        "summary": "Обновляет информацию о существующей заявке работодателя.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор заявки работодателя",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Обновлённая информация о заявке.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployerApplicationCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Заявка успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      }
    },
    "/api/EmployerApplication/{id}": {
      "get": {
        "tags": [
          "EmployerApplication"
        ],
        "summary": "Получает информацию о заявке работодателя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор заявки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка найдена и возвращена успешно.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployerApplicationDTO"
                }
              }
            }
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "EmployerApplication"
        ],
        "summary": "Удаляет заявку работодателя по идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор заявки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "204": {
            "description": "Заявка успешно удалена."
          },
          "404": {
            "description": "Заявка с указанным идентификатором не найдена."
          }
        }
      }
    },
...
    "/api/Query/applicants/employer-application/{employerApplicationId}": {
      "get": {
        "tags": [
          "Query"
        ],
        "summary": "Получает соискателей для заявки работодателя по идентификатору заявки.",
        "parameters": [
          {
            "name": "employerApplicationId",
            "in": "path",
            "description": "Идентификатор заявки работодателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список соискателей успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicantsForEmployerApplicationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicantsForEmployerApplicationDTO": {
        "type": "object",
        "properties": {
          "applicantId": {
            "type": "integer",
            "description": "Соискатель",
            "format": "int32"
          },
          "employerAppId": {
            "type": "integer",
            "description": "Заявка работодателя",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO описывающее соискателя и заявку работодателя"
      },
...
      "EmployerApplicationCreateDTO": {
        "type": "object",
        "properties": {
          "submissionDate": {
            "type": "string",
            "description": "Дата подачи",
            "format": "date-time"
          },
          "employerId": {
            "type": "integer",
            "description": "Работодатель",
            "format": "int32"
          },
          "positionId": {
            "type": "integer",
            "description": "Должность",
            "format": "int32"
          },
          "requirements": {
            "type": "string",
            "description": "Требование",
            "nullable": true
          },
          "offeredSalary": {
            "type": "integer",
            "description": "Предлагаемый уровень зарплаты",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания заявки работодателя"
      },
      "EmployerApplicationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор заявки работодателя",
            "format": "int32"
          },
          "submissionDate": {
            "type": "string",
            "description": "Дата подачи",
            "format": "date-time"
          },
          "employerId": {
            "type": "integer",
            "description": "Работодатель",
            "format": "int32"
          },
          "positionId": {
            "type": "integer",
            "description": "Должность",
            "format": "int32"
          },
          "requirements": {
            "type": "string",
            "description": "Требование",
            "nullable": true
          },
          "offeredSalary": {
            "type": "integer",
            "description": "Предлагаемый уровень зарплаты",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO заявки работодателя"
      },
      "EmployerCreateDTO": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "description": "Название компании",
            "nullable": true
          },
          "contactPersonName": {
            "type": "string",
            "description": "ФИО контактного лица",
            "nullable": true
          },
          "companyNumber": {
            "type": "string",
            "description": "Телефон работодателя",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания работодателя"
      },
      "EmployerDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор работадателя",
            "format": "int32"
          },
          "companyName": {
            "type": "string",
            "description": "Название компании",
            "nullable": true
          },
          "contactPersonName": {
            "type": "string",
            "description": "ФИО контактного лица",
            "nullable": true
          },
          "companyNumber": {
            "type": "string",
            "description": "Телефон работодателя",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO работодателя"
      },
...
    }
  }
}